requirements.txt
python-fasthtml
uvicorn
openai
dotenv
watchdog
requests
